add login and checking the code from the SMS

+++ src/main/java/com/example/graduateWork/repository/UsersRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface UsersRepository extends JpaRepository<Users, Long> {
-    Users findByPhoneNumber(int phoneNumber);
+    Users findByPhoneNumber(Long phoneNumber);
 
 }
